Add extra integration test and rename some
