+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/subdir0/go.mod
