+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/subdir0/subdir1/subsubdir1/add.go
@@ -1,0 +1,5 @@
+package subsubdir1
+
+func Add(a, b int) int {
+	return a + b
+}
