+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/subdir0/subdir2/add.go
@@ -1,0 +1,5 @@
+package subdir2
+
+func Add(a, b int) int {
+	return a + b
+}
