+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/subdir0/test.go
@@ -1,0 +1,14 @@
+package test
+
+import (
+	"test/subdir2"
+
+	"golang.org/x/net/ipv4"
+)
+
+func test() {
+
+	header := ipv4.Header{}
+	header.Version = subdir2.Add(2, 2)
+
+}
