+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/test.expected
@@ -1,0 +1,4 @@
+| Extraction failed in subdir0/subdir1/test.go with error cannot find package "subdir1/subsubdir1" in any of:\n\t/usr/local/Cellar/go/1.20.5/libexec/src/subdir1/subsubdir1 (from $GOROOT)\n\t/Users/owen-mc/go/src/subdir1/subsubdir1 (from $GOPATH) | 2 |
+| Extraction failed in subdir0/subdir1/test.go with error could not import subdir1/subsubdir1 (invalid package name: "") | 2 |
+| Extraction failed in subdir0/test.go with error cannot find package "test/subdir2" in any of:\n\t/usr/local/Cellar/go/1.20.5/libexec/src/test/subdir2 (from $GOROOT)\n\t/Users/owen-mc/go/src/test/subdir2 (from $GOPATH) | 2 |
+| Extraction failed in subdir0/test.go with error could not import test/subdir2 (invalid package name: "") | 2 |
