+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-none-in-root/test.py
