+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-one-in-root/go.mod
@@ -1,0 +1,5 @@
+go 1.14
+
+require golang.org/x/net v0.0.0-20200505041828-1ed23360d12c
+
+module main
