+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-one-in-root/subdir1/test.go
