+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-one-in-root/test.py
@@ -1,0 +1,8 @@
+import sys
+
+from create_database_utils import *
+from diagnostics_test_utils import *
+
+run_codeql_database_create([], lang="go")
+
+check_diagnostics()
