+++ go/ql/integration-tests/all-platforms/go/two-go-mods-nested-one-in-root/test.ql
@@ -1,0 +1,6 @@
+import go
+import semmle.go.DiagnosticsReporting
+
+from string msg, int sev
+where reportableDiagnostics(_, msg, sev)
+select msg, sev
