+++ go/ql/integration-tests/all-platforms/go/two-go-mods-not-nested/subdir1/go.mod
