+++ go/ql/integration-tests/all-platforms/go/two-go-mods-not-nested/subdir2/test.go
