--- go/ql/integration-tests/all-platforms/go/two-go-mods-one-in-root/test.expected
@@ -1,2 +1,0 @@
-| Extraction failed in test.go with error could not import subdir2 (invalid package name: "") | 2 |
-| Extraction failed in test.go with error package subdir2 is not in GOROOT (/usr/local/Cellar/go/1.20.5/libexec/src/subdir2) | 2 |
